bubble icon and sepparate overrunRight component
