+++ src/components/shared/BubbleIcon.tsx
@@ -1,0 +1,3 @@
+const BubbleIcon = ({ value }: { value: string }) => <div className="bubble-icon">{value}</div>;
+
+export default BubbleIcon;
