+++ src/components/shared/index.ts
@@ -15,3 +15,4 @@ export { default as FormFieldLabel } from './FormFieldLabel';
 export { default as ListField } from './ListField';
 export { default as HashTags } from './HashTagsForm';
 export { default as Toolbar } from './Toolbar';
+export { default as BubbleIcon } from './BubbleIcon';
